+++ packages/moonstone/CHANGELOG.md
@@ -7,9 +7,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - `moonstone/ContextualPopup`'s new API `positionContextualPopup`
+- `moonstone/Scroller` prop `onUpdate`
 
 ### Fixed
 
+- `moonstone/Scrollable` to update scroll properly on pointer click
 - `moonstone/TooltipDecorator` to prevent unnecessary re-renders when losing focus
 
 ## [2.0.0-beta.8] - 2018-06-25