+++ packages/moonstone/CHANGELOG.md
@@ -99,6 +99,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Slider` to read when knob gains focus or for a change in value
 - `moonstone/Scroller` to not cut off Expandables when scrollbar appears
 - `moonstone/VideoPlayer` to correctly read out when play button is pressed
+- `moonstone/Divider` to always use a fixed height, regardless of locale
 
 ## [2.0.0-beta.4] - 2018-05-21