+++ packages/moonstone/CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
+- `moonstone/MoonstoneDecorator` config property `disableFullscreen` to exclude `enact-fit` className
 - `moonstone/Scroller` prop `onUpdate`
 
 ### Fixed